Final commit hopefully, ripped out the attempted modular parser because it wasn't realistically achievable right now, the amount of variation from website to website would mean that the parser would need to be so complex in its configuration that it would be easier and faster to write a parser by hand for each website, which kind of defeats the point of having one in the first place.
The GUI itself was cleaned up and so was the main loop. The horizontal slider that used to determine the number of proxies to be scraped/generated was replaced by a pin wheel to allow the window to be moved while still retaining functionality. All but three of the tabs were removed, but the base functionality remains the same for everything but parsing the content, which admittedly worked like garbage anyway.